refactor: extract lobby into base lobby room and freeplay subclass
for code re-use with forthcoming tournament and education lobbies
