+++ client/src/views/Lobby.vue
@@ -24,7 +24,7 @@ import { Client, Room } from "colyseus.js";
 import { applyLobbyResponses } from "@port-of-mars/client/api/lobby/response";
 import { LobbyRequestAPI } from "@port-of-mars/client/api/lobby/request";
 import { AccountAPI } from "@port-of-mars/client/api/account/request";
-import { LOBBY_NAME } from "@port-of-mars/shared/lobby";
+import { FREE_PLAY_LOBBY_NAME } from "@port-of-mars/shared/lobby";
 import { GAME_PAGE, CONSENT_PAGE, MANUAL_PAGE } from "@port-of-mars/shared/routes";
 import { Constants } from "@port-of-mars/shared/settings";
 import { url } from "@port-of-mars/client/util";
@@ -76,7 +76,7 @@ export default class Lobby extends Vue {
   }
 
   async createRoom() {
-    const room = await this.$client.create(LOBBY_NAME);
+    const room = await this.$client.create(FREE_PLAY_LOBBY_NAME);
     await this.connectAndRedirect(room);
   }
