+++ server/src/services/admin.ts
@@ -1,6 +1,6 @@
 import { matchMaker } from "colyseus";
 import { GameRoom } from "@port-of-mars/server/rooms/game";
-import { LobbyRoom } from "@port-of-mars/server/rooms/lobby";
+import { FreePlayLobbyRoom } from "@port-of-mars/server/rooms/lobby/freeplay";
 import { ChatReport, ModerationAction } from "@port-of-mars/server/entity";
 import { BaseService } from "@port-of-mars/server/services/db";
 import {
@@ -43,7 +43,8 @@ export class AdminService extends BaseService {
   }
 
   async getLobbyData(): Promise<any> {
-    const lobby = await matchMaker.findOneRoomAvailable(LobbyRoom.NAME, {});
+    // FIXME: need to look at all lobby rooms (tournament)
+    const lobby = await matchMaker.findOneRoomAvailable(FreePlayLobbyRoom.NAME, {});
     return lobby ?? {};
   }
