+++ shared/src/lobby/index.ts
@@ -1,4 +1,6 @@
 export * from "./requests";
 export * from "./responses";
 
-export const LOBBY_NAME = "lobby";
+export const BASE_LOBBY_NAME = "base_lobby";
+export const FREE_PLAY_LOBBY_NAME = "free_play_lobby";
+export const TOURNAMENT_LOBBY_NAME = "tournament_lobby";
